+++ src/UI/TransactionData/TransactionData.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode } from 'react';
+import React, { ReactNode, useState } from 'react';
 import { IconProp } from '@fortawesome/fontawesome-svg-core';
 import classNames from 'classnames';
 
@@ -9,6 +9,7 @@ import { decodePart } from 'utils/decoders/decodePart';
 import { getUnHighlightedDataFieldParts } from 'utils/transactions/getUnHighlightedDataFieldParts';
 
 import { WithClassnameType } from '../types';
+import { TransactionDataDecode } from './components';
 
 const allOccurences = (sourceStr: string, searchStr: string) => {
   const occurrences = [...sourceStr.matchAll(new RegExp(searchStr, 'gi'))].map(
@@ -19,6 +20,7 @@ const allOccurences = (sourceStr: string, searchStr: string) => {
 };
 
 export interface TransactionDataPropsType extends WithClassnameType {
+  customCopyIcon?: IconProp;
   data: string;
   highlight?: string;
   innerTransactionDataClasses?: {
@@ -27,30 +29,33 @@ export interface TransactionDataPropsType extends WithClassnameType {
   };
   isScCall?: boolean;
   label?: ReactNode;
-  transactionIndex: number;
   showCopyButton?: boolean;
-  customCopyIcon?: IconProp;
+  showDataDecode?: boolean;
+  transactionIndex: number;
 }
 
 const TransactionDataComponent = ({
   className = 'dapp-transaction-data',
-  showCopyButton = true,
+  customCopyIcon,
   data,
   globalStyles,
   highlight,
   innerTransactionDataClasses,
   isScCall,
   label,
-  transactionIndex,
-  customCopyIcon,
-  styles
+  showCopyButton = true,
+  styles,
+  showDataDecode,
+  transactionIndex
 }: TransactionDataPropsType & WithStylesImportType) => {
+  const [decodedData, setDecodedData] = useState(data);
+
   const {
     transactionDataInputLabelClassName,
     transactionDataInputValueClassName
   } = innerTransactionDataClasses || {};
 
-  let output = <>{data}</>;
+  let output = <>{decodedData}</>;
 
   const [encodedScCall, ...remainingDataFields] =
     highlight && isScCall ? highlight.split('@') : [];
@@ -67,9 +72,17 @@ const TransactionDataComponent = ({
     element.scrollIntoView();
   };
 
+  const handleDecode = (decoded: string) => {
+    setDecodedData(decoded);
+  };
+
+  const handleDecodeError = () => {
+    setDecodedData(data);
+  };
+
   if (showHighlight) {
     switch (true) {
-      case data.startsWith(highlight): {
+      case decodedData.startsWith(highlight): {
         const [, rest] = data.split(highlight);
 
         output = (
@@ -101,7 +114,7 @@ const TransactionDataComponent = ({
         const { start, end } = getUnHighlightedDataFieldParts({
           occurrences,
           transactionIndex,
-          data,
+          data: data,
           highlight
         });
 
@@ -164,14 +177,21 @@ const TransactionDataComponent = ({
       )}
 
       <div className={classNames(styles?.transactionData, className)}>
-        <span
+        <div
           className={classNames(
             styles?.transactionDataLabel,
             transactionDataInputLabelClassName
           )}
         >
           {label ?? 'Data'}
-        </span>
+          {showDataDecode && (
+            <TransactionDataDecode
+              data={data}
+              onDecode={handleDecode}
+              onDecodeError={handleDecodeError}
+            />
+          )}
+        </div>
 
         <div className={styles?.transactionDataValueWrapper}>
           <div
@@ -182,13 +202,13 @@ const TransactionDataComponent = ({
             )}
           >
             <span className={styles?.transactionDataValueText}>
-              {data ? output : N_A}
+              {decodedData ? output : N_A}
             </span>
 
-            {data && showCopyButton && (
+            {decodedData && showCopyButton && (
               <CopyButton
                 copyIcon={customCopyIcon}
-                text={data}
+                text={decodedData}
                 className={styles?.transactionDataValueCopy}
               />
             )}