+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added transaction data decoder](https://github.com/multiversx/mx-sdk-dapp/pull/1311)
+
 ## [[v3.0.14](https://github.com/multiversx/mx-sdk-dapp/pull/1315)] - 2024-11-25
 
 - [Updated default metamaskSnapWalletAddress](https://github.com/multiversx/mx-sdk-dapp/pull/1314)